# docs/openapi-pos-mvp.yaml
openapi: 3.0.3
...
info:
  title: BoothPOS API
  version: 1.4.0
  description: |
    API internal untuk BoothPOS — sistem POS berbasis event untuk toko
    merchandise yang berjualan online maupun di bazaar/event. Berjalan di
    localhost pada perangkat operasional toko, tanpa akses publik,
    sejalan dengan model lisensi sekali bayar dan instalasi lokal per toko.

    **Autentikasi** — token bearer. Token diperoleh dari `/auth/login`
    dan dikirim pada header `Authorization: Bearer <token>`.

    **Format uang** — seluruh nilai uang berupa string desimal dua angka
    di belakang koma, contoh `"150000.00"`. String dipakai agar tidak ada
    galat pembulatan floating point di sisi klien.

    **Zona waktu** — seluruh timestamp dalam format ISO 8601 dengan
    offset waktu lokal perangkat.
servers:
  - url: http://localhost:8000/api/v1
    description: Instalasi lokal pada laptop operasional
...
security:
  - bearerAuth: []

# =====================================================================
# PATHS
# =====================================================================
paths:
...
  /reports/sales:
    get:
      tags: [Report]
      summary: Laporan penjualan
      parameters:
        - { name: event_id, in: query, schema: { type: integer } }
        - { name: date_from, in: query, schema: { type: string, format: date } }
        - { name: date_to, in: query, schema: { type: string, format: date } }
        - { name: group_by, in: query, schema: { type: string, enum: [product, category, artist, day, event], default: product } }
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalesReport' }
  /reports/profit:
    get:
      tags: [Report]
      summary: Laporan modal dan keuntungan
      description: |
        Laba kotor dihitung dari nilai snapshot pada `order_items`, bukan
        harga master saat laporan dibuka. Laba bersih mengurangkan
        `event_cost`. Hanya peran owner dan admin.
      parameters:
        - { name: event_id, in: query, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfitReport' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/artist-profit:
    get:
      tags: [Report]
      summary: Laporan modal dan laba kotor per artist (F9.5)
      description: |
        Untuk setiap artist dengan minimal satu `order_items` 'completed'
        pada event ini: `total_sales` (SUM `line_total`), `modal` (SUM
        `cost_price * qty`, snapshot yang sama dipakai `GET /reports/profit`
        — bukan modal BOM hidup, lihat docblock `ReportController::
        artistProfit()`), dan `gross_profit = total_sales - modal`.

        `event_cost` TIDAK dikurangkan di laporan ini sama sekali (sudah
        diperhitungkan terpisah pada `GET /reports/profit`, F9.3) — sengaja
        tidak ada field itu pada respons ini. Hanya peran owner dan admin.
      parameters:
        - { name: event_id, in: query, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  event: { $ref: '#/components/schemas/Event' }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        artist_id: { type: integer }
                        artist_name: { type: string }
                        total_sales: { $ref: '#/components/schemas/Money' }
                        modal: { $ref: '#/components/schemas/Money' }
                        gross_profit: { $ref: '#/components/schemas/Money' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/purchases:
    get:
      tags: [Report]
      summary: Laporan pembelian (006-purchase-order-and-ops, US9)
      description: |
        Sumbernya `purchase_orders` (Eloquent, sudah HasDataMode — bukan
        `DB::table()` mentah seperti /reports/sales), bukan `order_items`.
        Tidak diskop event. Hanya peran owner dan admin.
      parameters:
        - { name: vendor_id, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: string, enum: [draft, ordered, received, paid, cancelled] } }
        - { name: date_from, in: query, schema: { type: string, format: date } }
        - { name: date_to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  rows:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: integer }
                        po_number: { type: string }
                        vendor_name: { type: string }
                        status: { type: string }
                        created_at: { type: string, format: date-time }
                        total_amount: { $ref: '#/components/schemas/Money' }
                  totals:
                    type: object
                    properties:
                      po_count: { type: integer }
                      total_amount: { $ref: '#/components/schemas/Money' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/stock-by-artist:
    get:
      tags: [Report]
      summary: Laporan stok per artist (006-purchase-order-and-ops, US10)
      description: |
        Sumbernya `product_variants` (sudah HasDataMode via join produk),
        dijumlahkan `current_stock` per artist. Tidak diskop event. Hanya
        peran owner dan admin.
      parameters:
        - { name: artist_id, in: query, schema: { type: integer } }
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        artist_id: { type: integer }
                        artist_name: { type: string }
                        variant_count: { type: integer }
                        total_stock: { type: integer }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/artist-settlements:
    get:
      tags: [Report]
      summary: Rekap hasil per artist
      description: |
        Angka dihitung ulang dari `order_items` setiap kali dibaca, tanpa
        potongan komisi.

        Daftar memuat **seluruh artist aktif**, termasuk yang belum punya
        penjualan sama sekali di event ini (seluruh nilai uang `"0.00"`,
        `total_units` 0, `status` `unpaid`, `id` `null`). Artist nonaktif
        hanya ikut bila punya baris settlement di event ini.
      parameters:
        - { name: event_id, in: query, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  event: { $ref: '#/components/schemas/Event' }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ArtistSettlement' }
  /reports/artist-settlements/{artist}/transactions:
    parameters:
      - name: artist
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Report]
      summary: Drill-down transaksi yang menyusun rekap satu artist (F11.6)
      description: |
        Daftar order 'completed' pada event ini yang memuat minimal satu
        item milik artist ini, dikelompokkan per order. Karena satu order
        booth ini bisa berisi item dari BEBERAPA artist sekaligus, `items[]`
        pada setiap baris HANYA memuat item milik artist yang diminta —
        bukan seluruh isi order tersebut. Hanya peran owner dan admin.
      parameters:
        - { name: event_id, in: query, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  event: { $ref: '#/components/schemas/Event' }
                  artist:
                    type: object
                    properties:
                      id: { type: integer }
                      name: { type: string }
                  transactions:
                    type: array
                    items:
                      type: object
                      properties:
                        order_id: { type: integer }
                        order_number: { type: string }
                        created_at: { type: string, format: date-time, nullable: true }
                        items:
                          type: array
                          items:
                            type: object
                            properties:
                              sku: { type: string }
                              name: { type: string }
                              qty: { type: integer }
                              line_total: { $ref: '#/components/schemas/Money' }
                        order_total_for_artist:
                          $ref: '#/components/schemas/Money'
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/artist-settlements/{id}/payment:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Report]
      summary: Catat pembayaran ke artist
      description: Menandai berapa yang sudah diserahkan ke artist. Tidak mengubah `total_sales`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: string, example: "1250000.00" }
                notes: { type: string, nullable: true }
      responses:
        '200':
          description: Diperbarui
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArtistSettlement' }
  /reports/{report}/export:
    parameters:
      - name: report
        in: path
        required: true
        schema: { type: string, enum: [sales, profit, artist-settlements, artist-profit] }
    get:
      tags: [Report]
      summary: Ekspor laporan ke Excel
      description: |
        Parameter filter sama dengan endpoint laporan terkait. Ekspor tidak
        menyertakan kolom kontak pelanggan.

        `report=artist-settlements` (F11.6) sengaja berbeda dari tiga
        laporan lain: berkasnya berisi DUA sheet, "Rekap" (bentuk lama,
        tidak berubah) dan "Detail Transaksi" (baru — satu baris flat per
        item per order, lintas semua artist di event ini: artist_name,
        order_number, date, item_name, qty, line_total).
      parameters:
        - { name: event_id, in: query, schema: { type: integer } }
      responses:
        '200':
          description: Berkas xlsx
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }

  # ------------------------------------------------------------------
  # VENDOR, BAHAN BAKU, DAN BOM (ditambahkan pasca-MVP, 2026-09-01)
  #
  # CATATAN CAKUPAN — bukan kebangkitan "vendor management" / "materials/
  # production" yang dicoret PRD §10.2: cakupan di sini sengaja lebih
  # sempit (tidak ada purchase order, tidak ada penjadwalan produksi) dan
  # tidak berkorespondensi dengan nomor F- manapun. Lihat catatan
  # bertanggal di CLAUDE.md/README.md/PRD §10.2.
  # ------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  responses:
...
    Forbidden:
      description: Peran tidak memiliki izin
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        message: { type: string }
        code: { type: string, nullable: true, example: INSUFFICIENT_STOCK }
...
    Money:
      type: string
      pattern: '^-?\d+\.\d{2}$'
      example: "150000.00"

    # BUG YANG DITEMUKAN & DIPERBAIKI — skema ini sebelumnya juga bernama
    # `User` sama seperti skema entitas CRUD pengguna di bawah (lihat
    # bagian "Pengguna & Peran"), yang ditambahkan belakangan oleh alur
    # kerja terpisah. Dua definisi top-level dengan kunci sama di YAML
    # membuat definisi kedua diam-diam menimpa yang pertama pada parser
    # yang taat spesifikasi — endpoint /auth/login dan /auth/me akan
    # terdokumentasi memakai skema yang salah (bentuk objek role, bukan
    # nama+menu_keys) tanpa galat apa pun yang terlihat. Diberi nama
    # sendiri karena memang bentuknya berbeda dari skema entitas User
    # (role di sini nama tampilan saja + menu_keys hasil resolusi sesi,
    # bukan objek Role penuh).
...
    Event:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        location: { type: string, nullable: true }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        status: { type: string, enum: [draft, active, closed, cancelled] }
        event_cost: { $ref: '#/components/schemas/Money' }
        notes: { type: string, nullable: true }
...
    SalesReport:
      type: object
      description: |
        Task 1 — `group_label`, `rows[].entity_id`, dan `transactions`
        ditambahkan supaya KPI `totals.order_count` dan tabel `rows`
        (agregat per group_by) tidak lagi terlihat "berkontradiksi": dua
        transaksi berbeda yang membeli produk yang sama menghasilkan dua
        entri `transactions` tapi hanya satu baris `rows` — keduanya
        sama-sama benar, cuma menjawab pertanyaan berbeda.
      properties:
        event: { $ref: '#/components/schemas/Event' }
        group_by: { type: string, enum: [product, category, artist, day, event] }
        group_label:
          type: string
          description: |
            Label bahasa Indonesia siap-tampil untuk kolom pengelompokan
            (mis. "Produk"/"Kategori"/"Artist"/"Tanggal") — supaya
            frontend tidak perlu hardcode pemetaan group_by -> label.
          example: Produk
        totals:
          type: object
          properties:
            order_count: { type: integer }
            unit_count: { type: integer }
            gross_sales: { $ref: '#/components/schemas/Money' }
            discount_total: { $ref: '#/components/schemas/Money' }
            net_sales: { $ref: '#/components/schemas/Money' }
        rows:
          type: array
          items:
            type: object
            properties:
              entity_id:
                type: integer
                nullable: true
                description: |
                  product_id/category_id/artist_id sesuai group_by, untuk
                  ditaut ke halaman detail entitas (Task 2). `null` untuk
                  group_by=day, yang tidak menunjuk entitas apa pun.
              label: { type: string, description: "Nama produk, kategori, artist, atau tanggal sesuai group_by" }
              unit_count: { type: integer }
              amount: { $ref: '#/components/schemas/Money' }
        transactions:
          type: array
          description: |
            Satu baris = satu order berstatus completed pada rentang/event
            yang difilter — daftar transaksi individual yang sebelumnya
            tidak tersedia lewat endpoint ini. `id` disertakan supaya
            frontend bisa memanggil GET /orders/{id}/receipt untuk struk
            transaksi manapun di daftar ini (Task 3).
          items:
            type: object
            properties:
              id: { type: integer }
              order_number: { type: string }
              created_at: { type: string, format: date-time }
              cashier_name: { type: string, nullable: true }
              customer_name:
                type: string
                nullable: true
                description: |
                  F10.6 — nama pelanggan bila order dikaitkan ke satu
                  (`orders.customer_id` nullable — walk-in tidak punya
                  data pelanggan). Disediakan untuk pencarian di
                  frontend atas array `transactions[]` yang sudah
                  diambil (nomor transaksi/nama pelanggan/nama kasir);
                  tidak ada parameter pencarian sisi server yang baru.
              item_count: { type: integer }
              total_amount: { $ref: '#/components/schemas/Money' }
    ProfitReport:
      type: object
      properties:
        event: { $ref: '#/components/schemas/Event' }
        revenue: { $ref: '#/components/schemas/Money' }
        cost_of_goods: { $ref: '#/components/schemas/Money' }
        gross_profit: { $ref: '#/components/schemas/Money' }
        event_cost: { $ref: '#/components/schemas/Money' }
        net_profit: { $ref: '#/components/schemas/Money' }
    ArtistSettlement:
      type: object
      properties:
        id:
          type: integer
          nullable: true
          description: |
            ID baris `artist_settlements` — dipakai sebagai `{id}` pada
            `POST /reports/artist-settlements/{id}/payment`. Bernilai `null`
            untuk artist aktif yang belum punya penjualan di event ini
            (barisnya belum ada, dan memang tidak ada yang perlu dibayar).
            Pakai `artist_id` sebagai kunci baris di tabel UI.
        artist_id: { type: integer }
        artist_name: { type: string }
        total_sales: { $ref: '#/components/schemas/Money' }
        total_units: { type: integer }
        deduction: { $ref: '#/components/schemas/Money' }
        payable_amount: { $ref: '#/components/schemas/Money' }
        paid_amount: { $ref: '#/components/schemas/Money' }
        outstanding: { $ref: '#/components/schemas/Money' }
        status: { type: string, enum: [unpaid, partial, paid] }
        calculated_at: { type: string, format: date-time, nullable: true }
...